move util funcitons to util module
